+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Recommendation: for ease of reading, use the following order:
 ## [0.2.0] - 2025-07-23
 ### Added
 - Added a configurable interval for querying project changes during cyclic indexing.
+- API: `GET /system/state`: Added logging of applied operations.
 ### Changed
 - Improved iterative indexing.